MenuGen/requirements.txt
streamlit
huggingface_hub
diffusers
transformers
torch
Pillow
pytesseract
accelerate
opencv-pythont
huggingface_hub
diffusers
transformers
# torch
Pillow
pytesseract
accelerate
